+++ src/connectors/translation/DeeplTranslator.php
@@ -39,5 +39,4 @@ class DeeplTranslator implements TranslatorInterface
 
         return $altText;
     }
-
 }