+++ src/events/RegisterGeneratorsEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace fork\altify\events;
+
+use yii\base\Event;
+
+class RegisterGeneratorsEvent extends Event
+{
+    public array $generators = [];
+}